Struct dleq_vrf::keys::SecretKey

pub struct SecretKey<K: AffineRepr> { /* private fields */ }

Seceret key consisting of a scalar and a secret nonce seed.

Implementations

impl<K: AffineRepr> SecretKey<K>

pub fn from_xof(thin: ThinVrf<K>, xof: impl XofReader) -> Self

Generate an “unbiased” SecretKey from a user supplied XofReader.

pub fn from_seed(thin: ThinVrf<K>, seed: &[u8; 32]) -> Self

Generate a SecretKey from a 32 byte seed.

pub fn ephemeral(thin: ThinVrf<K>) -> Self

Generate an ephemeral SecretKey with system randomness.

pub fn as_publickey(&self) -> &PublicKey<K>

Reference the PublicKey corresponding to this SecretKey.

pub fn to_public(&self) -> PublicKey<K>

Clone the PublicKey corresponding to this SecretKey.

pub fn set_rng_for_test_vectors(&mut self)

pub fn witness(&self, t: &Transcript, label: impl AsLabel) -> Reader

impl<K: AffineRepr> SecretKey<K>

pub fn vrf_preout<H>(&self, input: &VrfInput<H>) -> VrfPreOut<H>where H: AffineRepr<ScalarField = K::ScalarField>,

Compute VRF pre-output from secret key and input.

pub fn vrf_inout<I, H>(&self, input: I) -> VrfInOut<H>where I: IntoVrfInput<H>, H: AffineRepr<ScalarField = K::ScalarField>,

Compute VRF pre-output paired with input from secret key and some VRF input, like a Transcript or a VrfInput.

As the arkworks hash-to-curve infrastructure looks complex, we employ arkworks’ simpler UniformRand here, which uses shitty try and increment. We strongly recommend you use a better hash-to-curve manually.

impl<K: AffineRepr> SecretKey<K>

pub fn sign_thin_vrf( &self, t: impl IntoTranscript, ios: &[VrfInOut<K>] ) -> Signature<ThinVrf<K>>

Sign thin VRF signature

If ios = &[] this reduces to a Schnorr signature.

impl<K: AffineRepr> SecretKey<K>

pub fn new_secret_blinding<const B: usize>( &self, t: &Transcript ) -> SecretBlinding<K, B>