Struct dleq_vrf::PedersenVrf

pub struct PedersenVrf<K, H = K, const B: usize = 1>where
    K: AffineRepr,
    H: AffineRepr<ScalarField = K::ScalarField>,{ /* private fields */ }

Pedersen VRF flavor

Implementations

impl<K, H, const B: usize> PedersenVrf<K, H, B>where K: AffineRepr, H: AffineRepr<ScalarField = K::ScalarField>,

pub fn new(keying_base: K, blinding_bases: [K; B]) -> PedersenVrf<K, H, B>

pub fn compute_blinded_publickey( &self, public: &PublicKey<K>, secret_blindings: &SecretBlinding<K, B> ) -> KeyCommitment<K>

impl<K, H, const B: usize> PedersenVrf<K, H, B>where K: AffineRepr, H: AffineRepr<ScalarField = K::ScalarField>,

pub fn sign_pedersen_vrf( &self, t: impl IntoTranscript, ios: &[VrfInOut<H>], secret_blinding: Option<SecretBlinding<K, B>>, secret: &SecretKey<K> ) -> (Signature<PedersenVrf<K, H, B>>, SecretBlinding<K, B>)

Sign Pedersen VRF signature

We create the secret blinding unless the user supplies one.

pub fn sign_non_batchable_pedersen_vrf( &self, t: impl IntoTranscript, ios: &[VrfInOut<H>], secret_blinding: Option<SecretBlinding<K, B>>, secret: &SecretKey<K> ) -> (NonBatchable<PedersenVrf<K, H, B>>, SecretBlinding<K, B>)

Sign a non-batchable Pedersen VRF signature

Non-batchable Pedersen VRF signatures resemble EC VRF, especially if B=1, except they contain the public key. We suggest thin VRF instead though for the EC VRF case when H=K and B=1.

We envision non-batchable VRF signatures being useful for proofs that public keys agree aross curves match, so when H != K but B=1.

We create the secret blinding unless the user supplies one.

impl<K, H, const B: usize> PedersenVrf<K, H, B>where K: AffineRepr, H: AffineRepr<ScalarField = K::ScalarField>,

pub fn verify_pedersen_vrf<'a>( &self, t: impl IntoTranscript, ios: &'a [VrfInOut<H>], signature: &Signature<PedersenVrf<K, H, B>> ) -> SignatureResult<&'a [VrfInOut<H>]>

Verify Pedersen VRF signature

pub fn verify_non_batchable_pedersen_vrf<'a>( &self, t: impl IntoTranscript, ios: &'a [VrfInOut<H>], signature: &NonBatchable<PedersenVrf<K, H, B>> ) -> SignatureResult<&'a [VrfInOut<H>]>

Verify Pedersen VRF signature