Trait dleq_vrf::traits::EcVrfSigner

pub trait EcVrfSigner: Borrow<Self::Secret> {
    type V: EcVrfVerifier;
    type Error;
    type Secret: EcVrfSecret<<Self::V as EcVrfVerifier>::H>;

    // Required method
    fn vrf_sign_detached(
        &self,
        t: impl IntoTranscript,
        ios: &[EcVrfInOut<Self::V>]
    ) -> Result<EcVrfProof<Self::V>, Self::Error>;

    // Provided methods
    fn vrf_sign<const N: usize>(
        &self,
        t: impl IntoTranscript,
        ios: &[EcVrfInOut<Self::V>; N]
    ) -> Result<VrfSignature<Self::V, N>, Self::Error> { ... }
    fn vrf_sign_one<I, T, F>(
        &self,
        input: I,
        check: F
    ) -> Result<VrfSignature<Self::V, 1>, Self::Error>
       where I: IntoVrfInput<<Self::V as EcVrfVerifier>::H>,
             T: IntoTranscript,
             F: FnMut(&EcVrfInOut<Self::V>) -> Result<T, Self::Error> { ... }
    fn vrf_sign_vec(
        &self,
        t: impl IntoTranscript,
        ios: &[EcVrfInOut<Self::V>]
    ) -> Result<VrfSignatureVec<Self::V>, Self::Error> { ... }
}

VRF signer, includes the secret key, but sometimes the ring opening too.

We do not favor pseduo-convenience methods like schnorrkel’s sign_extra_after_check. We’ve found too many VRF protocols need multiple input-output pairs, making convenience methods inconvenient. In practice, one should typically invoke EcVrfSecret::vrf_inout seperately for each input, run buisness logic upon the VrfInOuts they return, and then pass whatever VrfInOuts require signing to vrf_sign*.

We do provide an analogous vrf_sign_one method for pedagogy, but we also provide convenience methods to handle multiple input groupings instead of doing multiple single input flavors.

Inherent methods and other traits being used here: IntoTranscript, vrf::{VrfInOut, VrfPreOut}

Required Associated Types

type V: EcVrfVerifier

Verifier type

We only fetch V::H and V::VrfProof via this, so if your verifier contains lifetimes then approximate them by 'static.

type Error

Signer failures, usually ! but otherwise for ring VRFs

type Secret: EcVrfSecret<<Self::V as EcVrfVerifier>::H>

Actual secret key type, possible Self for regular VRFs but not for ring VRFs or simlar.

Required Methods

fn vrf_sign_detached( &self, t: impl IntoTranscript, ios: &[EcVrfInOut<Self::V>] ) -> Result<EcVrfProof<Self::V>, Self::Error>

Provided Methods

fn vrf_sign<const N: usize>( &self, t: impl IntoTranscript, ios: &[EcVrfInOut<Self::V>; N] ) -> Result<VrfSignature<Self::V, N>, Self::Error>

VRF signature for a fixed number of input-output pairs

fn vrf_sign_one<I, T, F>( &self, input: I, check: F ) -> Result<VrfSignature<Self::V, 1>, Self::Error>where I: IntoVrfInput<<Self::V as EcVrfVerifier>::H>, T: IntoTranscript, F: FnMut(&EcVrfInOut<Self::V>) -> Result<T, Self::Error>,

VRF signature for a single input-output pair

We provide this analog of schnorrkel’s sign_extra_after_check more for pedagogy than for convenience. It demonstrates choosing whether we sign the VRF, and what else we sign in its transcript, after examining the VRF output.

fn vrf_sign_vec( &self, t: impl IntoTranscript, ios: &[EcVrfInOut<Self::V>] ) -> Result<VrfSignatureVec<Self::V>, Self::Error>

VRF signature for a variable number of input-output pairs.